% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cal_all_tools}
\alias{cal_all_tools}
\title{calculation seven tools}
\usage{
cal_all_tools(
  counts,
  gSets_path,
  tools = c("AUCell", "pagoda2", "fscLVM", "Vision", "ROMA", "GSVA", "ssGSEA", "plage",
    "zscore"),
  filter = F,
  normalize = c("log", "CLR", "RC", "scran", "sctransform", "none"),
  n_cores = 3,
  rand_seed = 123
)
}
\arguments{
\item{counts}{gene expresion matrix, rows are genes and cols are cells}

\item{gSets_path}{patwhays/gene sets in clasical GMT format}

\item{filter}{whether filtering for genes expressed in less than 5% cells}

\item{normalize}{normalization method}

\item{tool}{select PAS tools}
}
\description{
parameters `counts`,  `gSets_path` and `gSets` are consistent across all tools
}
\details{
other parameters see details of tools
}
