% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calculate_PAS}
\alias{calculate_PAS}
\title{calculation Pathway Activity Score}
\usage{
calculate_PAS(
  counts,
  tool,
  species = "none",
  pathway = "none",
  gmt_file = "none",
  filter = F,
  normalize = "sctransform",
  n_cores = 3,
  rand_seed = 123
)
}
\arguments{
\item{counts}{gene expresion matrix, rows are genes and cols are cells}

\item{tool}{tool name}

\item{species}{species; human or mouse}

\item{pathway}{abbreviation for pathway database}

\item{gmt_file}{pathways in GMT format}

\item{filter}{whether filtering for genes expressed in less than 5 percent cells}

\item{normalize}{normalization method}
}
\description{
parameters `counts`,  `gSets_path` and `gSets` are consistent across all tools
}
\details{
other parameters see details of tools
}
\examples{
counts = load_counts()
score = calculate_PAS(counts, 'AUCell', 'human','kegg')
}
