% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prepare_vis}
\alias{prepare_vis}
\title{preparation of visulization object}
\usage{
prepare_vis(pas_score, n_pcs = 10, resolution = 0.3)
}
\arguments{
\item{pas_score}{a matrix of PAS}
}
\description{
preparation of visulization object
}
